% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battinghistory.R
\docType{data}
\name{battinghistory}
\alias{battinghistory}
\title{Baseball Batting History Data}
\format{
140 obs. of  27 variables:
\describe{
\item{Year}{season}
\item{Tms}{number of teams}
\item{N.Bat}{number of players}
\item{BatAge}{batter's average age}
\item{R}{runs scored}
\item{G}{games played}
\item{PA}{plate appearances}
\item{AB}{at-bats}
\item{H}{hits}
\item{X2B}{doubles}
\item{X3B}{triples}
\item{HR}{home runs}
\item{RBI}{runs batted in}
\item{SB}{stolen bases}
\item{CS}{number caught stealing}
\item{BB}{walks}
\item{SO}{strikeouts}
\item{BA}{batting average}
\item{OBP}{on-base percentage}
\item{SLG}{slugging percentage}
\item{OPS}{OBP plus SLG}
\item{TB}{total bases}
\item{GDP}{ground into double plays}
\item{HBP}{hit by pitches}
\item{SH}{sacrifice hits}
\item{SF}{sacrifice flies}
\item{IBB}{intentional walks}
}
}
\source{
baseball-reference.com.
}
\usage{
battinghistory
}
\description{
Major League Baseball data on batting; number of hits, doubles, home runs by season.
The data was extracted from baseball-reference.com website.
}
\note{
{This version of the data is sorted in ascending order of Year.
There are missing values, especially in early years.}
}
\keyword{datasets}
