% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bball.men.R
\docType{data}
\name{bball.men}
\alias{bball.men}
\title{Men's NCAA Basketball Data}
\format{
77 obs. of  20 variables:
\describe{
\item{Season}{season}
\item{Teams}{number of teams}
\item{G}{average number of games played}
\item{FG}{average number of field goals}
\item{FGA}{average number of field goal attempts}
\item{\verb{FG\%}}{field goal percentage}
\item{3P}{average number of three pointers}
\item{3PA}{average number of three point attempts}
\item{3P\%}{three-point percenrage}
\item{FT}{average number of free throws}
\item{FTA}{average number of free throw attempts}
\item{\verb{FT\%}}{free-throw percentage}
\item{TRB}{average number of total rebounds}
\item{AST}{average number of assists}
\item{STL}{average number of steals}
\item{BLK}{average number of blocks}
\item{TOV}{average number of turnovers}
\item{PF}{average number of personal fouls}
\item{PTS}{average number of points scored}
\item{Year}{Year season started}
}
}
\source{
Sports Reference
}
\usage{
bball.men
}
\description{
Description: Game averages for NCAA men basketball
}
\details{
The data is from Sports Reference
https://www.sports-reference.com/cbb/seasons/game-averages.html
}
\keyword{datasets}
