% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgsu.R
\docType{data}
\name{bgsu}
\alias{bgsu}
\title{BGSU Enrollment}
\format{
Data frame of selected BGSU enrollment data: 16 obs. of 2 variables
\describe{
\item{Year}{Year.}
\item{Enrollment}{Enrollment.}
}
}
\source{
J. Albert
}
\usage{
bgsu
}
\description{
BGSU Enrollment
}
\keyword{datasets}
