% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainsize.R
\docType{data}
\name{brainsize}
\alias{brainsize}
\title{Brain Size and Intelligence Data}
\format{
40 obs. of  7 variables:
\describe{
\item{Gender}{Male or Female.}
\item{FSIQ}{Full Scale IQ scores based on four Wechsler (1981) subtests.}
\item{VIQ}{Verbal IQ scores based on four Wechsler (1981) subtests.}
\item{PIQ}{Performance IQ scores based on four Wechsler (1981) subtests.}
\item{Weight}{Body weight in pounds.}
\item{Height}{Height in inches.}
\item{MRI_Count}{total pixel count from the 18 MRI scans.}
}
}
\source{
Willerman et al (1991).
}
\usage{
brainsize
}
\description{
Data from a study comparing brain size and intelligence.
}
\note{
{There are missing values in Weight (2) and Height (1).}
}
\keyword{datasets}
