% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college.R
\docType{data}
\name{college}
\alias{college}
\title{College Rating Data}
\format{
260 obs. of  11 variables:
\describe{
\item{School}{Name of Institution.}
\item{Enrollment}{Enrollment of Institution.}
\item{Tier}{Ranking in tiers 1, 2, 3, 4.}
\item{Retention}{Pct. of freshmen who return the following year}
\item{Grad.rate}{Pct. of freshmen who graduate in six years}
\item{Pct.20}{Pct. of classes with 20 or fewer students}
\item{Pct.50}{Pct. of classes with 50 or fewer students}
\item{Full.time}{Pct. of faculty hired full-time}
\item{Top.10}{Pct. of incoming students who were in top 10\% of high school class}
\item{Accept.rate}{Acceptance rate of students who apply}
\item{Alumni.giving}{Pct. of alumni who contribute financially}
}
}
\source{
US News and World Report "America's Best Colleges" 2009 report, National Universities.
}
\usage{
college
}
\description{
College Rating Data
}
\note{
{There are missing values.}
}
\keyword{datasets}
