% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPUspeed.R
\docType{data}
\name{CPUspeed}
\alias{CPUspeed}
\title{CPU Speed Data}
\format{
27 obs. of 6 variables:
\describe{
\item{year}{calendar year}
\item{month}{month}
\item{day}{day}
\item{time}{time in years}
\item{speed}{Max IA-32 Speed (GHz)}
\item{log10speed}{logarithm base 10 of speed}
}
}
\usage{
CPUspeed
}
\description{
Maximum Intel CPU speed vs time from 1994 through 2004.
}
\keyword{datasets}
