% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime.bigcity.R
\docType{data}
\name{crime.bigcity}
\alias{crime.bigcity}
\title{Hartigan's City Crime Data}
\format{
16 obs. of  8 variables:
\describe{
\item{city}{name of city (character)}
\item{murder}{murder rate}
\item{rape}{rape rate}
\item{robbery}{robbery rate}
\item{assault}{assault rate}
\item{burglary}{burglary rate}
\item{larceny}{larceny rate}
\item{auto}{auto crime rate}
}
}
\source{
{
United Sates Statistical Abstracts (1970).
\url{https://people.sc.fsu.edu/~jburkardt/datasets/hartigan/file03.txt}}
}
\usage{
crime.bigcity
}
\description{
Number of crimes per 100,000 population, as of 1970 for 16 large cities in the US. Table 1.1 in Chapter 1 of Hartigan (1975).
All variables are numeric except city, which is character type.
}
\references{
{
Hartigan, J. A. (1975). Clustering Algorithms, John Wiley, New York.}
}
\keyword{datasets}
