% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decathlon.R
\docType{data}
\name{Decathlon}
\alias{Decathlon}
\title{Decathlon Data}
\format{
41 obs. of  14 variables:
\describe{
\item{Athlete}{name of athlete}
\item{100m}{100 meter}
\item{Long.jump}{long jump}
\item{Shot.put}{shotput}
\item{High.jump}{high jump}
\item{400m}{400 meter}
\item{110m.hurdle}{110 meter hurdles}
\item{Discus}{discus}
\item{Pole.vault}{pole vault}
\item{Javelin}{javelin}
\item{1500m}{1500m run}
\item{Rank}{rank}
\item{Points}{points}
\item{Competition}{'Decastar' or 'OlympicG'}
}
}
\source{
{ This tibble is essentially the same data as
in "decathlon.csv"
from Kaggle and other sources, but with changes in variable names
and format of athlete names.}
}
\usage{
Decathlon
}
\description{
This data records performance of athletes in two decathlon competitions, the
2004 Olympic Games and 2004 Decastar. Performance in each of the ten events
of the decathlon, rank, and points are recorded for each of the two
competitions. Except for "Athlete" (character) and "Competition"
(factor), all variables are numeric. There are no missing
values.
}
\note{
{
In the FactoMineR 'decathlon' version of the data,
athlete name was used as a row label. Row labels must be unique,
so there were upper case names for
Decastar and title case names for OlympicG.
In this version 'Decathlon' is a tibble, and athlete name is a
variable, so there is no need for unique values. The names in Decathlon
are all in title case, and the event can be
identified by the factor Competition. }
}
\keyword{datasets}
