% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EtruscanItalian.R
\docType{data}
\name{EtruscanItalian}
\alias{EtruscanItalian}
\title{Etruscan-Italian Data}
\format{
154 obs. of 2 variables:
\describe{
\item{x}{skull measurement}
\item{group}{character: Etruscan or Italian}
}
}
\usage{
EtruscanItalian
}
\description{
This data provides measurements of ancient Etruscan skulls and modern Italian
skulls.
}
\keyword{datasets}
