% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubble.R
\docType{data}
\name{hubble}
\alias{hubble}
\title{Hubble Space Telescope Data}
\format{
24 obs. of 3 variables:
\describe{
\item{Galaxy}{A label to identify the galaxy (a factor)}
\item{Velocity}{Relative velocity in kilometers per second}
\item{Distance}{Distance in Mega parsecs}
}
}
\source{
Freedman et al. 2001. The Astrophysical Journal 553:47-72: Tables 4 and 5.
}
\usage{
hubble
}
\description{
Distances and velocities measured for 24 galaxies containing
Cepheid stars to measure the Hubble constant.
}
\references{
Freedman et al. (2001) Final results from the Hubble space telescope key project to measure the Hubble constant. The Astrophysical Journal (553), 47-72.
Wood, S.N. (2017) Generalized Additive Models: An Introduction with R. CRC
}
\keyword{datasets}
