% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc.marathon.R
\docType{data}
\name{nyc.marathon}
\alias{nyc.marathon}
\title{New York City Marathon Data}
\format{
276 obs. of  3 variables:
\describe{
\item{Gender}{female or male}
\item{Minutes}{Time of runner in minutes}
\item{Age}{Age of runner}
}
}
\usage{
nyc.marathon
}
\description{
Gender, age, and completion time (in minutes) for 276 people who
completed the 2010 New York City Marathon.
}
\keyword{datasets}
