% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peanuts.R
\docType{data}
\name{peanuts}
\alias{peanuts}
\title{Peanuts Aflatoxin Data}
\format{
34 obs. of 2 variables:
\describe{
\item{Percent}{percentage of non-contaminated peanuts in the batch}
\item{Aflatoxin}{average level of aflatoxin in parts per billion}
}
}
\source{
Hand et al. (1994)
}
\usage{
peanuts
}
\description{
The peanuts data records levels of a toxin (aflatoxin) in batches of
peanuts.
}
\keyword{datasets}
