% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison.R
\docType{data}
\name{poison}
\alias{poison}
\title{Poison Survival Data}
\format{
48 obs. of 3 variables:
\describe{
\item{Time}{survival time in units of 10 hours}
\item{Poison}{poison: I, II, III}
\item{Treatment}{treatment: A, B, C, D}
}
}
\source{
Box, G. E. P., Hunter, W. G. and Hunter, J. S. (1978), Statistics for
Experimenters: An Introduction to Design, Data Analysis, and Model Building,
Wiley, New York.
}
\usage{
poison
}
\description{
Survival times in units of 10 hours for animals exposed to different poisons.
}
\keyword{datasets}
