% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\docType{data}
\name{rounding}
\alias{rounding}
\title{Rounding First Base Data}
\format{
66 obs. of 3 variables:
\describe{
\item{times}{time}
\item{method}{factor with 3 levels: NarrowAngle, RoundOut, WideAngle}
\item{block}{player ID (integer)}
}
}
\source{
Hollander and Wolfe (1999) Table 7.1, page 274.
}
\usage{
rounding
}
\description{
Times required to round first base for 22 baseball players using three
styles: rounding out, a narrow angle and a wide angle.
The goal is to determine if the method of rounding first base has a
significant effect on times to round first base.
}
\keyword{datasets}
