% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiRstv.R
\docType{data}
\name{SiRstv}
\alias{SiRstv}
\title{NIST SiRstv Data}
\format{
25 obs. of 2 variables:
\describe{
\item{Instrument}{replicate}
\item{Resistance}{resistance}
}
}
\source{
https://www.itl.nist.gov/div898/strd/anova/SiRstv.html
}
\usage{
SiRstv
}
\description{
Measurements of bulk resistivity of silicon wafers made at NIST with 5
probing instruments on each of 5 days.
}
\details{
https://www.itl.nist.gov/div898/strd/anova/SiRstv_info.html
}
\references{
NIST Standard Reference Datasets: https://www.itl.nist.gov/div898/strd/index.html
}
\keyword{datasets}
