% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfall.R
\docType{data}
\name{snowfall}
\alias{snowfall}
\title{Buffalo and Cleveland Snowfall Data}
\format{
41 obs. of 3 variables:
\describe{
\item{SEASON}{character: winter season identified by years}
\item{Cleveland}{Cleveland snowfall}
\item{Buffalo}{Buffalo snowfall}
}
}
\usage{
snowfall
}
\description{
Total snowfall in inches for the cities Buffalo and Cleveland for the seasons
1968-69 through 2008-09.
}
\keyword{datasets}
