% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgrades.R
\docType{data}
\name{statgrades}
\alias{statgrades}
\title{Statistics Grades}
\format{
23 obs. of 7 variables:
\describe{
\item{ID}{Student ID; integer 1:23}
\item{Exam1}{Percent grade on Exam 1}
\item{Exam2}{Percent grade on Exam 2}
\item{HW}{Percent grade on homework}
\item{Final}{Percent grade on Final Exam}
\item{Major}{Major coded 1, 2, 3}
\item{Group}{Group coded 1, 2}
}
}
\usage{
statgrades
}
\description{
Grades from an undergraduate statistics class at BGSU.
}
\keyword{datasets}
