% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinIQ.R
\docType{data}
\name{twinIQ}
\alias{twinIQ}
\title{Twins IQ Data}
\format{
Data frame of Burt's IQ data for twins: 27 obs. of 3 variables
\describe{
\item{Foster}{IQ of twin raised with foster parents.}
\item{Biological}{IQ of twin raised with biological parents.}
\item{Social}{Social class of biological parents (high, low, middle)}
}
}
\source{
Burt, C. (1966). The genetic estimation of differences in intelligence:
A study of monozygotic twins reared together and apart. Br. J. Psych., 57, 147-153.
Data is provided in R packages faraway and UsingR.
}
\usage{
twinIQ
}
\description{
Twins IQ Data
}
\keyword{datasets}
