% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utley2006.R
\docType{data}
\name{utley2006}
\alias{utley2006}
\title{Chase Utley's Hitting Data for 2006}
\format{
160 obs. of 6 variables:
\describe{
\item{Game}{game}
\item{Date}{date}
\item{PA}{plate appearances}
\item{AB}{at-bats}
\item{R}{home runs}
\item{H}{hits}
}
}
\source{
J. Albert
}
\usage{
utley2006
}
\description{
Chase Utley's Hitting Data for 2006
}
\details{
During the 2006 baseball season, Chase Utley of the Philadelphia Phillies
had a hitting streak of 35 games, which is one of the best hitting streaks
in baseball history.
}
\keyword{datasets}
