% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasterunup.R
\docType{data}
\name{wasterunup}
\alias{wasterunup}
\title{Waste Run-up Data}
\format{
22 obs. of 5 variables:
\describe{
\item{PT1}{weekly percentage waste of cloth for Plant 1}
\item{PT2}{weekly percentage waste of cloth for Plant 2}
\item{PT3}{weekly percentage waste of cloth for Plant 3}
\item{PT4}{weekly percentage waste of cloth for Plant 4}
\item{PT5}{weekly percentage waste of cloth for Plant 5}
}
}
\usage{
wasterunup
}
\description{
The 'Waste Run-up' data (Koopmans 1987, p. 86) reports weekly percentage
waste of cloth by five different supplier plants of Levi-Strauss, relative
to cutting from a computer pattern.
}
\note{
There are missing values.
}
\keyword{datasets}
