% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhits.R
\docType{data}
\name{webhits}
\alias{webhits}
\title{Webpage Hits Data}
\format{
35 obs. of 2 variables:
\describe{
\item{Week}{Week number}
\item{Hits}{Number of web hits}
}
}
\source{
J. Albert
}
\usage{
webhits
}
\description{
The number of daily visits to the author's website was obtained using
Google Analytics. The data is summarized by week.
}
\keyword{datasets}
